+++ src/sprite/Scale.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include "Scale.h"
 
-trippin::Scale::Scale(std::string name, double multiplier) : name(std::move(name)), multiplier(multiplier) {
+trippin::Scale::Scale(std::string name, int multiplier) : name(std::move(name)), multiplier(multiplier) {
 
 }
 
@@ -9,6 +9,6 @@ const std::string &trippin::Scale::getName() const {
     return name;
 }
 
-double trippin::Scale::getMultiplier() const {
+int trippin::Scale::getMultiplier() const {
     return multiplier;
 }
